requirements.txt
einops==0.4.1
yerbamate.egg==info
pytorch_lightning==1.7.5
torchmetrics==0.9.3
tensorflow_gpu==2.10.0
ipdb==0.13.9
transformers==4.22.0
torch==1.12.1+cu116
torchvision==0.13.1+cu116
